requirements.txt
asgiref==3.5.2
astroid==2.11.5
certifi==2022.5.18.1
charset-normalizer==2.0.12
dill==0.3.5.1
Django==3.2.13
django-rest-framework==0.1.0
djangorestframework==3.13.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Naked==0.1.31
ordering==0.2.0
platformdirs==2.5.2
pylint==2.13.9
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
status==0.2.5
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1
